requirements.txt
playwright>=1.45.0
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.0.3
openai>=1.0.0
python-dotenv>=1.0.0